Move Poet::CLI into its own file
